Refactor memberFormatting.ts into memberData/common.ts
