+++ src/components/GridMemberList.astro
@@ -2,9 +2,10 @@
 // © 2024 Vlad-Stefan Harbuz <vlad@vlad.website>
 // SPDX-License-Identifier: Apache-2.0
 
+import { getMembers } from '../memberCollections.ts';
 import {
-  getMembers, filterInactiveMembers, sortMembersByDevs, fmtCurrency, fmtDevs, getDollarsPerDev
-} from '../memberFormatting.ts';
+  filterInactiveMembers, sortMembersByDevs, fmtCurrency, fmtDevs, getDollarsPerDev
+} from '../memberData/common.ts';
 import type { Map } from '../util.ts';
 import memberRoles from "../memberRoles.json"
