+++ src/components/MiniLeaderboard.astro
@@ -2,9 +2,8 @@
 // © 2024 Vlad-Stefan Harbuz <vlad@vlad.website>
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-  getMembers, filterInactiveMembers, sortMembersByDevs,
-} from '../memberFormatting.ts';
+import { getMembers } from '../memberCollections.ts';
+import { filterInactiveMembers, sortMembersByDevs } from '../memberData/common.ts';
 
 const N_TO_CHOOSE = 5;
 const members = sortMembersByDevs(filterInactiveMembers(await getMembers()));
