+++ src/components/TabularMemberList.astro
@@ -2,9 +2,10 @@
 // © 2024 Vlad-Stefan Harbuz <vlad@vlad.website>
 // SPDX-License-Identifier: Apache-2.0
 
+import { getMembers } from '../memberCollections.ts';
 import {
-  getMembers, filterInactiveMembers, getDollarsPerDev, fmtCurrency, sortMembersByDevs
-} from '../memberFormatting.ts';
+  filterInactiveMembers, getDollarsPerDev, fmtCurrency, sortMembersByDevs
+} from '../memberData/common.ts';
 
 const members = filterInactiveMembers(await getMembers());
 ---
