+++ src/components/TotalTally.astro
@@ -2,14 +2,13 @@
 // © 2024 Vlad-Stefan Harbuz <vlad@vlad.website>
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-  getGrandTotalRaised, fmtCurrency,
-} from '../memberFormatting.ts';
+import { getGrandTotalRaised, fmtCurrency } from '../memberData/common.ts';
+import { getMembers } from '../memberCollections.ts';
 ---
 
 <div class="total-tally">
   <p>Our members have paid maintainers</p>
-  <h2>{fmtCurrency(await getGrandTotalRaised())}</h2>
+  <h2>{fmtCurrency(getGrandTotalRaised(await getMembers()))}</h2>
   <p>over the last year.</p>
 </div>
