+++ src/pages/about.astro
@@ -8,9 +8,8 @@ import Button from "../components/Button.astro";
 import Layout from "../layouts/Layout.astro";
 import MiniLeaderboard from "../components/MiniLeaderboard.astro";
 import TextButton from "../components/TextButton.astro";
-import {
-  getMembers, filterInactiveMembers,
-} from '../memberFormatting.ts';
+import { filterInactiveMembers } from '../memberData/common.ts';
+import { getMembers } from '../memberCollections.ts';
 import TableOfContents from "../components/TableOfContents.astro";
 
 const members = filterInactiveMembers(await getMembers());
